[alerting] Move alerting outside of probes
