+++ internal/alerting/notifier/opsgenie/api_test.go
@@ -18,8 +18,8 @@ import (
 	"io"
 	"testing"
 
-	"github.com/cloudprober/cloudprober/probes/alerting/alertinfo"
-	configpb "github.com/cloudprober/cloudprober/probes/alerting/proto"
+	"github.com/cloudprober/cloudprober/internal/alerting/alertinfo"
+	configpb "github.com/cloudprober/cloudprober/internal/alerting/proto"
 	"github.com/stretchr/testify/assert"
 )
