+++ internal/alerting/notifier/pagerduty/event_v2.go
@@ -27,7 +27,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/cloudprober/cloudprober/probes/alerting/alertinfo"
+	"github.com/cloudprober/cloudprober/internal/alerting/alertinfo"
 )
 
 // EventV2Request is the data structure for a Pagerduty event.
