+++ internal/alerting/notifier/slack/slack.go
@@ -10,9 +10,9 @@ import (
 	"net/http"
 	"os"
 
+	"github.com/cloudprober/cloudprober/internal/alerting/alertinfo"
+	configpb "github.com/cloudprober/cloudprober/internal/alerting/proto"
 	"github.com/cloudprober/cloudprober/logger"
-	"github.com/cloudprober/cloudprober/probes/alerting/alertinfo"
-	configpb "github.com/cloudprober/cloudprober/probes/alerting/proto"
 )
 
 const (
