+++ probes/proto/config_proto_gen.cue
@@ -4,7 +4,7 @@ import (
 	"github.com/cloudprober/cloudprober/targets/proto"
 	proto_1 "github.com/cloudprober/cloudprober/metrics/proto"
 	proto_5 "github.com/cloudprober/cloudprober/validators/proto"
-	proto_A "github.com/cloudprober/cloudprober/probes/alerting/proto"
+	proto_A "github.com/cloudprober/cloudprober/internal/alerting/proto"
 	proto_8 "github.com/cloudprober/cloudprober/probes/ping/proto"
 	proto_E "github.com/cloudprober/cloudprober/probes/http/proto"
 	proto_B "github.com/cloudprober/cloudprober/probes/dns/proto"
