+++ web/web.go
@@ -26,8 +26,8 @@ import (
 	"github.com/cloudprober/cloudprober/common/httputils"
 	"github.com/cloudprober/cloudprober/config"
 	"github.com/cloudprober/cloudprober/config/runconfig"
+	"github.com/cloudprober/cloudprober/internal/alerting"
 	"github.com/cloudprober/cloudprober/probes"
-	"github.com/cloudprober/cloudprober/probes/alerting"
 	"github.com/cloudprober/cloudprober/servers"
 	"github.com/cloudprober/cloudprober/surfacers"
 	"github.com/cloudprober/cloudprober/web/resources"
